feat: allow specification of revisionHistoryLimit on certificates
Every time cert-manager renews a certificate it creates a new `CertificateRequest` resource for that renewal and, in cert-manager versions older than 1.18.0, old `CertificateRequest` resources are never cleaned up.  Version 1.18.0 and later keeps only the latest request by default, but in order to make older cert-manager delete old requests after a renewal it is necessary to specify an explicit `revisionHistoryLimit` on the `Certificate` resource.

This change sets a default history limit of 1, matching the default in cert-manager >=1.18.0, but allows this to be overridden via chart values.

Fixes #762

Signed-off-by: Ian Roberts <i.roberts@sheffield.ac.uk>